+++ test.go
@@ -112,7 +112,7 @@ const (
 	configFileName = "test/config.json"
 
 	// List of flavors for which a bootstrap Docker image is available.
-	flavors = "mysql80,percona80"
+	flavors = "mysql80,mysql84,percona80"
 )
 
 // Config is the overall object serialized in test/config.json.